split device methods
