+++ lib/appium_lib_core/device/screen_record.rb
@@ -10,6 +10,27 @@ module Appium
 
         METHOD = %w(POST PUT).freeze
 
+        def self.add_methods
+          Appium::Core::Device.add_endpoint_method(:stop_recording_screen) do
+            def stop_recording_screen(remote_path: nil, user: nil, pass: nil, method: 'PUT')
+              option = ::Appium::Core::Device::ScreenRecord.new(
+                remote_path: remote_path, user: user, pass: pass, method: method
+              ).upload_option
+
+              params = option.empty? ? {} : { options: option }
+
+              execute(:stop_recording_screen, {}, params)
+            end
+          end
+
+          Appium::Core::Device.add_endpoint_method(:stop_and_save_recording_screen) do
+            def stop_and_save_recording_screen(file_path)
+              base64data = execute(:stop_recording_screen, {}, {})
+              File.open(file_path, 'wb') { |f| f << Base64.decode64(base64data) }
+            end
+          end
+        end
+
         def initialize(remote_path: nil, user: nil, pass: nil, method: 'PUT', force_restart: nil)
           @upload_option = if remote_path.nil?
                              {}
